wip, tests passing
