+++ scio-test/src/test/scala/com/spotify/scio/coders/serializers/JTraversableSerializerTest.scala
@@ -14,7 +14,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.spotify.scio.coders
+package com.spotify.scio.coders.serializers
 
 import com.esotericsoftware.kryo.io.{Input, Output}
 import com.twitter.chill.{Kryo, KryoSerializer}
